Segragated comparison page into 2 componants
